# morphio-io/frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Morphio.io API",
		"description": "This API will blow your mind",
		"version": "1.0.0"
	},
	"paths": {
...
		"/logs/process-logs": {
			"post": {
				"tags": ["Logs", "Logs"],
				"summary": "Process Logs",
				"description": "Enqueue a log file for processing and return a job ID.",
				"operationId": "process_logs",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "anonymize",
						"in": "query",
						"required": false,
						"schema": {
							"type": "boolean",
							"description": "Anonymize sensitive data before processing",
							"default": false,
							"title": "Anonymize"
						},
						"description": "Anonymize sensitive data before processing"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_process_logs"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Log processing job enqueued successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_LogsProcessingResponse_"
								},
								"example": {
									"status": "success",
									"message": "Log processing job enqueued",
									"data": {
										"job_id": "123e4567-e89b-12d3-a456-426614174000",
										"status": "pending",
										"message": "Log processing job enqueued."
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/logs/logs-processing-status/{job_id}": {
			"get": {
				"tags": ["Logs", "Logs"],
				"summary": "Get Logs Processing Status Route",
				"description": "Retrieve the status of a log processing job.",
				"operationId": "get_logs_processing_status",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "job_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Job Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Log processing status retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_LogsProcessingStatusResponse_"
								},
								"example": {
									"status": "success",
									"message": "Log processing status retrieved",
									"data": {
										"job_id": "123e4567-e89b-12d3-a456-426614174000",
										"status": "completed",
										"progress": 100,
										"message": "Log summary generated and saved",
										"result": {
											"summary": "Summary content here",
											"title": "Log Summary",
											"content_id": 1,
											"user_id": 1,
											"template_id": "log-summary"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Job not found"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/logs/generate-splunk-config": {
			"post": {
				"tags": ["Logs", "Logs"],
				"summary": "Generate Splunk Config",
				"description": "Enqueue a log file sample for Splunk config generation and return a job ID.",
				"operationId": "generate_splunk_config",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "anonymize",
						"in": "query",
						"required": false,
						"schema": {
							"type": "boolean",
							"description": "Anonymize sensitive data before processing",
							"default": false,
							"title": "Anonymize"
						},
						"description": "Anonymize sensitive data before processing"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_generate_splunk_config"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Splunk config generation job enqueued successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_LogsProcessingResponse_"
								},
								"example": {
									"status": "success",
									"message": "Splunk config generation job enqueued",
									"data": {
										"job_id": "123e4567-e89b-12d3-a456-426614174000",
										"status": "pending",
										"message": "Splunk config generation job enqueued."
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/logs/config": {
			"get": {
				"tags": ["Logs", "Logs"],
				"summary": "Get Log File Configuration",
				"description": "Get log file configuration including allowed extensions.",
				"operationId": "get_log_config",
				"responses": {
					"200": {
						"description": "Log file configuration",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_dict_"
								},
								"example": {
									"status": "success",
									"message": "Log file configuration",
									"data": {
										"allowed_extensions": ["csv", "json", "log", "md", "txt"]
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"Body_generate_splunk_config": {
				"properties": {
					"log_file": {
						"type": "string",
						"format": "binary",
						"title": "Log File"
					}
				},
				"type": "object",
				"required": ["log_file"],
				"title": "Body_generate_splunk_config"
			},
			"Body_process_logs": {
				"properties": {
					"log_file": {
						"type": "string",
						"format": "binary",
						"title": "Log File"
					}
				},
				"type": "object",
				"required": ["log_file"],
				"title": "Body_process_logs"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"JobStatus": {
				"type": "string",
				"enum": [
					"pending",
					"processing",
					"completed",
					"failed",
					"cancelled",
					"not_found"
				],
				"title": "JobStatus"
			},
			"LogsProcessingResponse": {
				"properties": {
					"job_id": {
						"type": "string",
						"title": "Job Id"
					},
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Message"
					}
				},
				"type": "object",
				"required": ["job_id", "status"],
				"title": "LogsProcessingResponse"
			},
			"LogsProcessingStatusResponse": {
				"properties": {
					"job_id": {
						"type": "string",
						"title": "Job Id"
					},
					"status": {
						"$ref": "#/components/schemas/JobStatus"
					},
					"progress": {
						"type": "integer",
						"title": "Progress",
						"default": 0
					},
					"message": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Message"
					},
					"result": {
						"anyOf": [
							{
								"additionalProperties": true,
								"type": "object"
							},
							{
								"type": "null"
							}
						],
						"title": "Result"
					},
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Error"
					}
				},
				"type": "object",
				"required": ["job_id", "status"],
				"title": "LogsProcessingStatusResponse"
			},
...
			"ResponseModel_LogsProcessingResponse_": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/LogsProcessingResponse"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[LogsProcessingResponse]"
			},
			"ResponseModel_LogsProcessingStatusResponse_": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/LogsProcessingStatusResponse"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[LogsProcessingStatusResponse]"
			},
...
			"ResponseModel_dict_": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"additionalProperties": true,
								"type": "object"
							},
							{
								"type": "null"
							}
						],
						"title": "Data"
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[dict]"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
...
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": {
					"password": {
						"scopes": {},
						"tokenUrl": "/auth/login"
					}
				}
			}
		}
	}
}